Infinite map size, new maps, position checker
I added the ability for the program to read the map's size (width,height) before reading the map itself. In the main function of DungeonCrawler.java, it will display the map and ask for X and Y positions you would like to check. Edge of the map still produces out of bounds errors, and X and Y get counted from the upper left coordinate.