Lots of changes
Unimplemented ideas put in place for checking room data around the player. Solidified the role of the Player class.

Player: Stores X and Y position, as well as data for what items are collected. Has a checkPosition function that stores the map values for north, south, east, and west as variables, then calls map.printDirection with the data
DungeonMap: has checkNorth,south,etc and checkCurrent to gather information on the map at the certain positions. Also, printDirection, which takes in the data and prints a string based on the number data. This means that the old checkPos function is now obsolete but I do not have enough brainpower atm to remove it and implement a new solution in DungeonCrawler.
DungeonCrawler: has a new checkRoom function which is used for when you walk into a certain room. I made an example for walking into a room with a sword.