Big changes
Game works as an actual game! Compile DungeonCrawler.java to test it.
It can read maps, tell what the starting position is, and you can go through the dungeon.
TO DO: The game needs to have a way to quit!
We need to update the other dungeons.
We should start working on score system too!